Refactored some utils, and added a config.
